added failure check on cryptsetup

+++ blocks/filesystem/gpt_luks_passphrase_ext4.sh
@@ -45,8 +45,20 @@ sgdisk -c ${PARTITION_CRYPT_SWAP}:"${LABEL_SWAP}" ${DRIVE}
 sgdisk -c ${PARTITION_CRYPT_ROOT}:"${LABEL_ROOT}" ${DRIVE}
 
 # format LUKS on root
+
+# let cryptsetup handle password entry, exit after 3 successive failures
+_tries=0; _failed=true; while _failed; do
+_tries=$((_tries+1))
 cryptsetup --cipher=aes-xts-plain --verify-passphrase --key-size=512 luksFormat ${DRIVE}${PARTITION_CRYPT_ROOT}
+[ $_tries -gt 2 ] && exit;
+[ $? -eq 0 ] && _failed=false; done
+
+# let cryptsetup handle password entry, exit after 3 successive failures
+_tries=0; _failed=true; while _failed; do
+_tries=$((_tries+1))
 cryptsetup luksOpen ${DRIVE}${PARTITION_CRYPT_ROOT} ${LABEL_ROOT_CRYPT}
+[ $_tries -gt 2 ] && exit;
+[ $? -eq 0 ] && _failed=false; done
 
 # make filesystems
 mkfs.vfat ${DRIVE}${PARTITION_EFI_BOOT}
